requirements.txt
fastapi==0.111.0
psycopg2==2.9.9
sqlalchemy==2.0.30
bcrypt==3.2.0
passlib==1.7.4
pymysql==1.1.1
alembic==1.13.1
python-jose==3.3.0
pytest==8.2.2
pytest-asyncio==0.23.7